refactored explanations and split out different files
