+++ src/features/available-area/explanations.js
@@ -1,23 +1,23 @@
 /**
- * @import { AacContext, AacExplanations, ActionWithArea, CompatibilityCheckFn, DesignationOverlap, DesignationZone } from './available-area.d.js'
- * @import { CodeToString, EligibilityEntry } from './available-area.d.js'
+ * @import { AacContext, AacExplanations } from './available-area.d.js'
+ * @import { CodeToString } from './available-area.d.js'
  * @import { ExplanationSection } from './explanations.d.js'
- * @import { LandCover } from '~/src/features/parcel/parcel.d.js'
  */
 
-import { sqmToHaRounded } from '~/src/features/common/helpers/measurement.js'
-import { TARGET_SUFFIX } from './availableArea.js'
-
-/**
- * Strips the target suffix from an action code if present.
- * @param {string} code
- * @returns {string}
- */
-function stripTargetSuffix(code) {
-  return code.endsWith(TARGET_SUFFIX)
-    ? code.slice(0, -TARGET_SUFFIX.length)
-    : code
-}
+import { deriveExplanations } from './explanation-derivation.js'
+import {
+  buildApplicationSection,
+  buildAdjustedActionsSection,
+  buildLandCoversOnParcelSection,
+  buildDesignationAreasSection,
+  buildDesignationEligibilitySection,
+  buildEligibilitySection,
+  buildIncompatibilitySection,
+  buildAllocationsSection,
+  buildTargetAvailabilitySection,
+  buildResultSection,
+  buildStacksSection
+} from './explanation-sections.js'
 
 /**
  * Creates basic explanation section
@@ -226,576 +226,4 @@ function buildFeasibleSolutionSections(
   }
 
   return sections
-}
-
-/**
- * Derives structured explanations from the AAC context.
- * @param {AacContext} aacContext
- * @returns {AacExplanations}
- */
-function deriveExplanations(aacContext) {
-  const {
-    solution,
-    targetLabel,
-    existingActions,
-    landCoversForParcel,
-    eligibility,
-    cliques,
-    compatibilityCheckFn
-  } = aacContext
-
-  // Eligibility: which land covers each action can use
-  const eligibilityExplanation =
-    /** @type { Record<string, EligibilityEntry[]> } */ ({})
-  for (const [actionCode, indices] of eligibility) {
-    const displayCode = stripTargetSuffix(actionCode)
-    eligibilityExplanation[displayCode] = indices.map((lcIdx) => ({
-      landCoverIndex: lcIdx,
-      landCoverClassCode: landCoversForParcel[lcIdx].landCoverClassCode,
-      areaSqm: landCoversForParcel[lcIdx].areaSqm
-    }))
-  }
-
-  // Existing actions (always derived, even without a solution)
-  const adjustedActions = existingActions.map((a) => ({
-    actionCode: a.actionCode,
-    areaSqm: a.areaSqm
-  }))
-
-  // Incompatibility cliques (only those with 2+ members)
-  const incompatibilityCliques = cliques
-    .filter((c) => c.length >= 2)
-    .map((c) => c.map(stripTargetSuffix))
-
-  // No solution means no LP was run or LP was infeasible
-  if (!solution) {
-    const noSolutionTargetIndices = eligibility.get(targetLabel) ?? []
-    return {
-      eligibility: eligibilityExplanation,
-      adjustedActions,
-      incompatibilityCliques,
-      allocations: [],
-      targetAvailability: noSolutionTargetIndices.map((lcIdx) => ({
-        landCoverIndex: lcIdx,
-        totalAreaSqm: landCoversForParcel[lcIdx].areaSqm,
-        usedByExistingSqm: 0,
-        availableSqm: landCoversForParcel[lcIdx].areaSqm
-      })),
-      stacks: []
-    }
-  }
-
-  return {
-    eligibility: eligibilityExplanation,
-    adjustedActions,
-    incompatibilityCliques,
-    allocations: buildAllocationsFromSolution(
-      existingActions,
-      eligibility,
-      solution
-    ),
-    targetAvailability: buildTargetAvailabilityFromSolution(
-      eligibility,
-      targetLabel,
-      solution,
-      landCoversForParcel
-    ),
-    stacks: buildStacksFromSolution(
-      solution,
-      existingActions,
-      eligibility,
-      compatibilityCheckFn
-    )
-  }
-}
-
-/**
- * Builds target availability per-land-cover breakdown from the LP solution.
- * @param {Map<string, number[]>} eligibility
- * @param {string} targetLabel
- * @param {object} solution
- * @param {LandCover[]} landCoversForParcel
- * @returns {{landCoverIndex: number, totalAreaSqm: number, usedByExistingSqm: number, availableSqm: number}[]}
- */
-function buildTargetAvailabilityFromSolution(
-  eligibility,
-  targetLabel,
-  solution,
-  landCoversForParcel
-) {
-  const targetIndices = eligibility.get(targetLabel) ?? []
-  return targetIndices.map((lcIdx) => {
-    const availableSqm = solution[`t_${lcIdx}`] || 0
-    const totalAreaSqm = landCoversForParcel[lcIdx].areaSqm
-    return {
-      landCoverIndex: lcIdx,
-      totalAreaSqm,
-      usedByExistingSqm: totalAreaSqm - availableSqm,
-      availableSqm
-    }
-  })
-}
-
-/**
- * Builds flat allocations array from the LP solution.
- * @param {ActionWithArea[]} existingActions
- * @param {Map<string, number[]>} eligibility
- * @param {object} solution
- * @returns {{actionCode: string, landCoverIndex: number, areaSqm: number}[]}
- */
-function buildAllocationsFromSolution(existingActions, eligibility, solution) {
-  const allocations = []
-  for (const action of existingActions) {
-    const eligibleIndices = eligibility.get(action.actionCode) ?? []
-    for (const lcIdx of eligibleIndices) {
-      const value = solution[`x_${action.actionCode}_${lcIdx}`] || 0
-      if (value > 0.001) {
-        allocations.push({
-          actionCode: action.actionCode,
-          landCoverIndex: lcIdx,
-          areaSqm: value
-        })
-      }
-    }
-  }
-  return allocations
-}
-
-/**
- * Derives stacks from the LP solution by examining co-located actions.
- * Uses a "peeling" algorithm: on each land cover, repeatedly peel off the
- * thinnest compatible layer until all area is accounted for. This produces
- * stacks that partition the land cover — their areas sum to the total used
- * area rather than overlapping.
- * @param {object} solution - LP solver result
- * @param {ActionWithArea[]} existingActions
- * @param {Map<string, number[]>} eligibility
- * @param {CompatibilityCheckFn} compatibilityCheckFn
- * @returns {import('./available-area.d.js').Stack[]}
- */
-function buildStacksFromSolution(
-  solution,
-  existingActions,
-  eligibility,
-  compatibilityCheckFn
-) {
-  const allocationsByLc = buildAllocationsByLandCover(
-    existingActions,
-    eligibility,
-    solution
-  )
-  const stacks = []
-  let stackNumber = 1
-
-  for (const [lcIdx, actions] of allocationsByLc) {
-    const lcStacks = buildStacksForLandCover(actions, compatibilityCheckFn)
-    for (const stack of lcStacks) {
-      stacks.push({
-        stackNumber,
-        actionCodes: stack.actionCodes,
-        areaSqm: stack.areaSqm,
-        landCoverIndex: lcIdx
-      })
-      stackNumber++
-    }
-  }
-
-  return stacks
-}
-
-/**
- * Finds the action with the smallest remaining area in a map.
- * @param {Map<string, number>} remaining
- * @returns {{ code: string, area: number }}
- */
-function findSmallestAction(remaining) {
-  let smallestCode = /** @type {string} */ ('')
-  let smallestArea = Infinity
-  for (const [code, area] of remaining) {
-    if (area < smallestArea) {
-      smallestArea = area
-      smallestCode = code
-    }
-  }
-  return { code: smallestCode, area: smallestArea }
-}
-
-/**
- * Builds stacks for a single land cover by peeling off compatible layers.
- * Each iteration finds the action with the smallest remaining area, groups
- * it with all compatible actions, and peels off that area as a stack.
- * @param {{actionCode: string, areaSqm: number}[]} actions
- * @param {CompatibilityCheckFn} compatibilityCheckFn
- * @returns {{actionCodes: string[], areaSqm: number}[]}
- */
-function buildStacksForLandCover(actions, compatibilityCheckFn) {
-  /** @type {Map<string, number>} */
-  const remaining = new Map(actions.map((a) => [a.actionCode, a.areaSqm]))
-  const stacks = []
-
-  while (remaining.size > 0) {
-    const { code: smallestCode, area: smallestArea } =
-      findSmallestAction(remaining)
-
-    // Build a maximal compatible group containing the smallest action
-    const group = [smallestCode]
-    const candidates = [...remaining.keys()]
-      .filter((c) => c !== smallestCode)
-      .sort(
-        (a, b) =>
-          /** @type {number} */ (remaining.get(b)) -
-          /** @type {number} */ (remaining.get(a))
-      )
-
-    for (const candidate of candidates) {
-      if (group.every((g) => compatibilityCheckFn(g, candidate))) {
-        group.push(candidate)
-      }
-    }
-
-    stacks.push({ actionCodes: group, areaSqm: smallestArea })
-
-    // Subtract the peeled area from all group members
-    for (const code of group) {
-      const newArea = /** @type {number} */ (remaining.get(code)) - smallestArea
-      if (newArea <= 0.001) {
-        remaining.delete(code)
-      } else {
-        remaining.set(code, newArea)
-      }
-    }
-  }
-
-  // Sort by area descending so the largest stacks appear first
-  stacks.sort((a, b) => b.areaSqm - a.areaSqm)
-  return stacks
-}
-
-/**
- * @param {ActionWithArea[]} existingActions
- * @param {Map<string, number[]>} eligibility
- * @param {object} solution
- * @returns {Map<number, {actionCode: string, areaSqm: number}[]>}
- */
-function buildAllocationsByLandCover(existingActions, eligibility, solution) {
-  const allocations = new Map()
-  for (const action of existingActions) {
-    const eligibleIndices = eligibility.get(action.actionCode) ?? []
-    for (const lcIdx of eligibleIndices) {
-      const value = solution[`x_${action.actionCode}_${lcIdx}`] || 0
-      if (value <= 0.001) {
-        continue
-      }
-
-      if (!allocations.has(lcIdx)) {
-        allocations.set(lcIdx, [])
-      }
-      allocations
-        .get(lcIdx)
-        .push({ actionCode: action.actionCode, areaSqm: value })
-    }
-  }
-  return allocations
-}
-
-const ZONE_LABELS = {
-  neither: 'no designation',
-  sssi_only: 'SSSI only',
-  hf_only: 'HF only',
-  sssi_and_hf: 'SSSI and HF'
-}
-
-/**
- * Formats a land cover name with an optional designation zone label.
- * @param {string} lcName - The land cover display name
- * @param {DesignationZone} [zone] - The designation zone, if any
- * @returns {string}
- */
-function formatLcWithZone(lcName, zone) {
-  if (!zone || zone === 'neither') {
-    return lcName
-  }
-  return `${lcName} [${ZONE_LABELS[zone]}]`
-}
-
-/**
- * Builds the Application section listing the target action code.
- * @param {string} targetAction
- * @returns {ExplanationSection}
- */
-function buildApplicationSection(targetAction) {
-  return { title: 'Application', content: [`Target action: ${targetAction}`] }
-}
-
-/**
- * Builds the land covers on the parcel section.
- * @param {LandCover[]} landCovers
- * @param {CodeToString} landCoverToString
- * @returns {ExplanationSection}
- */
-function buildLandCoversOnParcelSection(landCovers, landCoverToString) {
-  const content = landCovers.map(
-    (lc) =>
-      `${landCoverToString(lc.landCoverClassCode)}: ${sqmToHaRounded(lc.areaSqm)} ha`
-  )
-  return { title: 'Land covers on the parcel', content }
-}
-
-/**
- * Builds a section showing the SSSI and HF designation areas per land cover.
- * @param {LandCover[]} originalLandCovers
- * @param {CodeToString} landCoverToString
- * @param {DesignationOverlap[]} [sssiOverlap]
- * @param {DesignationOverlap[]} [hfOverlap]
- * @param {DesignationOverlap[]} [sssiAndHfOverlap]
- * @returns {ExplanationSection}
- */
-function buildDesignationAreasSection(
-  originalLandCovers,
-  landCoverToString,
-  sssiOverlap,
-  hfOverlap,
-  sssiAndHfOverlap
-) {
-  const content = []
-  for (let i = 0; i < originalLandCovers.length; i++) {
-    const lcName = landCoverToString(originalLandCovers[i].landCoverClassCode)
-    const sssi = sssiOverlap?.[i]?.areaSqm ?? 0
-    const hf = hfOverlap?.[i]?.areaSqm ?? 0
-    const both = sssiAndHfOverlap?.[i]?.areaSqm ?? 0
-    if (sssi > 0 || hf > 0) {
-      const parts = []
-      parts.push(`${sqmToHaRounded(sssi)} ha SSSI`)
-      parts.push(`${sqmToHaRounded(hf)} ha HF`)
-      if (both > 0) {
-        parts.push(`${sqmToHaRounded(both)} ha SSSI and HF overlap`)
-      }
-      content.push(`${lcName}: ${parts.join(', ')}`)
-    }
-  }
-  if (content.length === 0) {
-    content.push('No SSSI or HF designations on this parcel')
-  }
-  return { title: 'Designation areas', content }
-}
-
-/**
- * Builds a section showing the SSSI/HF eligibility of each action.
- * @param {string} targetLabel
- * @param {ActionWithArea[]} existingActions
- * @param {{[actionCode: string]: boolean}} [sssiActionEligibility]
- * @param {{[actionCode: string]: boolean}} [hfActionEligibility]
- * @returns {ExplanationSection}
- */
-function buildDesignationEligibilitySection(
-  targetLabel,
-  existingActions,
-  sssiActionEligibility,
-  hfActionEligibility
-) {
-  const targetRealCode = stripTargetSuffix(targetLabel)
-
-  const allCodes = [targetRealCode, ...existingActions.map((a) => a.actionCode)]
-
-  const content = allCodes.map((code) => {
-    const sssi =
-      sssiActionEligibility?.[code] !== false
-        ? 'SSSI eligible'
-        : 'SSSI ineligible'
-    const hf =
-      hfActionEligibility?.[code] !== false ? 'HF eligible' : 'HF ineligible'
-    return `${code}: ${sssi}, ${hf}`
-  })
-
-  return { title: 'Designation eligibility per action', content }
-}
-
-/**
- * @param {AacExplanations} explanations
- * @param {CodeToString} landCoverToString
- * @param {DesignationZone[]} [designationZones]
- * @returns {ExplanationSection}
- */
-function buildEligibilitySection(
-  explanations,
-  landCoverToString,
-  designationZones
-) {
-  const content = []
-
-  for (const [actionCode, entries] of Object.entries(
-    explanations.eligibility
-  )) {
-    if (entries.length === 0) {
-      content.push(`${actionCode}: no eligible land covers on this parcel`)
-    } else {
-      const landCoverList = entries
-        .map((e) => {
-          const lcName = landCoverToString(e.landCoverClassCode)
-          const zone = designationZones?.[e.landCoverIndex]
-          const display = formatLcWithZone(lcName, zone)
-          return display
-        })
-        .join(', ')
-      content.push(`${actionCode}: ${landCoverList}`)
-    }
-  }
-
-  return {
-    title: designationZones
-      ? 'Eligible land covers per action (with designation zones)'
-      : 'Eligible land covers per action',
-    content
-  }
-}
-
-/**
- * @param {AacExplanations} explanations
- * @returns {ExplanationSection}
- */
-function buildAdjustedActionsSection(explanations) {
-  const content = explanations.adjustedActions.map(
-    (adj) => `${adj.actionCode}: ${sqmToHaRounded(adj.areaSqm)} ha`
-  )
-
-  return {
-    title: 'Existing actions',
-    content
-  }
-}
-
-/**
- * @param {AacExplanations} explanations
- * @returns {ExplanationSection}
- */
-function buildIncompatibilitySection(explanations) {
-  const content = explanations.incompatibilityCliques.map(
-    (clique) => `${clique.join(', ')} cannot share the same land`
-  )
-
-  return {
-    title: 'Incompatible action groups',
-    content
-  }
-}
-
-/**
- * @param {AacExplanations} explanations
- * @param {LandCover[]} landCoversForParcel
- * @param {CodeToString} landCoverToString
- * @param {DesignationZone[]} [designationZones]
- * @returns {ExplanationSection}
- */
-function buildAllocationsSection(
-  explanations,
-  landCoversForParcel,
-  landCoverToString,
-  designationZones
-) {
-  /** @type {Map<string, {actionCode: string, landCoverIndex: number, areaSqm: number}[]>} */
-  const byAction = new Map()
-  for (const alloc of explanations.allocations) {
-    if (!byAction.has(alloc.actionCode)) {
-      byAction.set(alloc.actionCode, [])
-    }
-    const actionAllocs = /** @type {typeof explanations.allocations} */ (
-      byAction.get(alloc.actionCode)
-    )
-    actionAllocs.push(alloc)
-  }
-
-  const content = []
-  for (const [actionCode, allocs] of byAction) {
-    const parts = allocs.map((a) => {
-      const lc = landCoversForParcel[a.landCoverIndex]
-      const lcName = landCoverToString(lc.landCoverClassCode)
-      const zone = designationZones?.[a.landCoverIndex]
-      return `${sqmToHaRounded(a.areaSqm)} ha on ${formatLcWithZone(lcName, zone)}`
-    })
-    content.push(`${actionCode}: ${parts.join(', ')}`)
-  }
-
-  return {
-    title: 'Optimal placement of existing actions',
-    content
-  }
-}
-
-/**
- * @param {AacExplanations} explanations
- * @param {string} targetAction
- * @param {LandCover[]} landCoversForParcel
- * @param {CodeToString} landCoverToString
- * @param {DesignationZone[]} [designationZones]
- * @returns {ExplanationSection}
- */
-function buildTargetAvailabilitySection(
-  explanations,
-  targetAction,
-  landCoversForParcel,
-  landCoverToString,
-  designationZones
-) {
-  const content = explanations.targetAvailability.map((ta) => {
-    const lc = landCoversForParcel[ta.landCoverIndex]
-    const lcName = landCoverToString(lc.landCoverClassCode)
-    const zone = designationZones?.[ta.landCoverIndex]
-    const display = formatLcWithZone(lcName, zone)
-    return `${display}: ${sqmToHaRounded(ta.totalAreaSqm)} ha total, ${sqmToHaRounded(ta.usedByExistingSqm)} ha used by existing actions, ${sqmToHaRounded(ta.availableSqm)} ha available`
-  })
-
-  return {
-    title: `Available land for ${targetAction}`,
-    content
-  }
-}
-
-/**
- * @param {string} targetAction
- * @param {number} availableAreaSqm
- * @param {number} totalValidLandCoverSqm
- * @returns {ExplanationSection}
- */
-function buildResultSection(
-  targetAction,
-  availableAreaSqm,
-  totalValidLandCoverSqm
-) {
-  return {
-    title: 'Result',
-    content: [
-      `Total eligible land cover for ${targetAction}: ${sqmToHaRounded(totalValidLandCoverSqm)} ha`,
-      `Maximum available area for ${targetAction}: ${sqmToHaRounded(availableAreaSqm)} ha`
-    ]
-  }
-}
-
-/**
- * @param {AacExplanations} explanations
- * @param {LandCover[]} landCoversForParcel
- * @param {CodeToString} landCoverToString
- * @param {DesignationZone[]} [designationZones]
- * @returns {ExplanationSection}
- */
-function buildStacksSection(
-  explanations,
-  landCoversForParcel,
-  landCoverToString,
-  designationZones
-) {
-  const content = explanations.stacks.map((stack) => {
-    let display = 'Land cover unknown'
-    if (stack.landCoverIndex !== undefined) {
-      const lc = landCoversForParcel[stack.landCoverIndex]
-      const lcName = landCoverToString(lc.landCoverClassCode)
-      const zone = designationZones?.[stack.landCoverIndex]
-      display = formatLcWithZone(lcName, zone)
-    }
-    return `Stack ${stack.stackNumber}: ${stack.actionCodes.join(' + ')} on ${display} (${sqmToHaRounded(stack.areaSqm)} ha)`
-  })
-
-  return {
-    title: 'Stacks',
-    content
-  }
 }
